// constants/error.constants.d.ts
declare const errorMessage: {
    NOTFOUND: {
        status: number;
        message: string;
    };
    LOGINERROR: {
        status: number;
        message: string;
    };
    UNAUTHORIZED: {
        status: number;
        message: string;
    };
    INVALIDDATA: {
        status: number;
        message: string;
    };
    EXISTS: {
        status: number;
        message: string;
    };
    OUTOFSTOCK: {
        status: number;
        message: string;
    };
    NOTBALANCE: {
        status: number;
        message: string;
    };
    EXHAUSTED: {
        status: number;
        message: string;
    };
    ALREADYTAKEN: {
        status: number;
        message: string;
    };
    IDEMPOTENCYKEYREUSED: {
        status: number;
        message: string;
    };
};
export { errorMessage };
//# sourceMappingURL=error.constants.d.ts.map